+++ TDS_Server/Manager/Logs/RestLogsManager.cs
@@ -3,6 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.Net;
 using System.Threading.Tasks;
+using TDS_Common.Enum;
 using TDS_Server.Enum;
 using TDS_Server.Manager.Player;
 using TDS_Server_DB.Entity;
@@ -16,7 +17,7 @@ namespace TDS_Server.Manager.Logs
             bool ipAddressParseWorked = IPAddress.TryParse(source?.Address, out IPAddress address);
             var log = new LogRests
             {
-                Type = (short)type,
+                Type = type,
                 Source = source?.GetEntity()?.Id ?? 0,
                 Ip = ipAddressParseWorked ? address : null,
                 Serial = saveipserial ? source?.Serial : null,
